Added basic, imperfect checks for TTS being correctly initialized.
There seem to be complications caused by how Android supports the calls TextToSpeech.Engine.ACTION_CHECK_TTS_DATA on newer versions of Android e.g. 4.3 devices. (see http://stackoverflow.com/questions/13558337/text-to-speech-not-working-in-android-4-2-jelly-bean/13633757#13633757 )

I'll continue to enhance the detection and initialization of the TTS engine, with the relevant voice(s) in future iterations of this project.